requirements.txt
adal==1.2.7
alembic==1.13.2
annotated-types==0.7.0
antlr4-python3-runtime==4.13.1
anyio==4.4.0
applicationinsights==0.11.10
argcomplete==3.3.0
azure-cli-telemetry==1.1.0
azure-common==1.1.28
azure-core==1.35.0
azure-cosmos==3.2.0
azure-identity==1.21.0
azure-keyvault-administration==4.4.0b2
azure-keyvault-certificates==4.7.0
azure-keyvault-keys>=4.9,<5.0
azure-keyvault-secrets==4.7.0
azure-mgmt-compute==34.0.0
azure-mgmt-core==1.5.0
azure-mgmt-datalake-nspkg==3.0.1
azure-mgmt-nspkg==3.0.2
azure-mgmt-resource==23.3.0
azure-mgmt-sqlvirtualmachine==1.0.0b5
azure-mgmt-storage==22.1.1
azure-mgmt-synapse==2.1.0b5
azure-mgmt-trafficmanager==1.0.0
azure-mgmt-web==8.0.0
azure-monitor-query==2.0.0
azure-multiapi-storage==1.2.0
azure-nspkg==3.0.2
azure-storage-blob==12.24.1
azure-storage-common==1.4.2
azure-synapse-accesscontrol==0.5.0
azure-synapse-artifacts==0.18.0
azure-synapse-managedprivateendpoints==0.4.0
azure-synapse-spark==0.2.0
bcrypt==4.1.3
blinker==1.8.2
boto3==1.39.15
cachelib==0.9.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==44.0.1
decorator==5.1.1
Deprecated==1.2.14
distro==1.9.0
fabric==3.2.2
Flask==2.0.1
Flask-Bcrypt==0.7.1
Flask-Caching==2.3.0
Flask-Limiter==3.8.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
google-auth==2.29.0
google-auth-oauthlib==1.2.0
greenlet==1.1.3.post0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
humanfriendly==10.0
idna==3.7
importlib_resources==6.4.5
invoke==2.2.0
isodate==0.6.1
itsdangerous==2.0.1
argon2-cffi==23.1.0
javaproperties==0.5.2
Jinja2==3.1.4
jmespath==1.0.1
jsondiff==2.0.0
knack==0.11.0
limits==3.13.0
Mako==1.3.5
Markdown==3.6
markdown-it-py==3.0.0
markdown2==2.4.13
MarkupSafe==2.1.5
mdurl==0.1.2
msal==1.30.0
msal-extensions==1.2.0
msrest==0.7.1
msrestazure==0.6.4
matplotlib==3.9.0
oauthlib==3.2.2
openai==1.62.0
ordered-set==4.1.0
packaging==24.0
paramiko==3.4.0
pathlib==1.0.1
pillow==10.4.0
pkginfo==1.10.0
portalocker==2.8.2
psutil==5.9.8
prowler==5.11.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycomposefile==0.0.31
pycparser==2.22
pydantic==2.9.2
pygame==2.6.0
PyGithub==2.5.0
Pygments==2.18.0
PyJWT==2.8.0
# pymsalruntime==0.14.2
PyNaCl==1.5.0
pyotp==2.9.0
pyreadline3==3.4.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
#pywin32==306
PyYAML==6.0.2
reportlab==4.2.2
requests==2.32.2
requests-oauthlib==2.0.0
rich==13.8.1
rsa==4.7.2
ruff==0.6.5
scp==0.13.6
semver==2.13.0
setuptools==75.1.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==1.4.23
sshtunnel==0.1.5
svgwrite==1.4.3
tabulate==0.9.0
templates==0.0.5
tqdm==4.66.4
Tree==0.2.4
typing_extensions==4.12.0
urllib3>=1.25.4,<1.27
waitress==3.0.0
websocket-client==1.3.3
Werkzeug==2.0.1
wrapt==1.16.0
WTForms==3.1.2
xmltodict==0.13.0
pymysql==1.1.0
rq~=2.1.0
































